blog auth

+++ blog/app/Http/Controllers/ArticleController.php
@@ -4,9 +4,17 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
+use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Auth;
 
 class ArticleController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware("auth")->except(["index", "detail"]);
+    }
+
     public function index()
     {
         $data = Article::latest()->paginate(5);
@@ -18,7 +26,11 @@ class ArticleController extends Controller
 
     public function add()
     {
-        return view("articles.add");
+        $categories = Category::all();
+
+        return view("articles.add", [
+            'categories' => $categories,
+        ]);
     }
 
     public function create()
@@ -37,6 +49,7 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = Auth::id();
         $article->save();
 
         return redirect("/articles");
@@ -54,8 +67,12 @@ class ArticleController extends Controller
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
+        
+        if(Gate::allows("delete-article", $article)) {
+            $article->delete();
+            return redirect("/articles")->with("info", "An article is deleted");
+        }
 
-        return redirect("/articles")->with("info", "An article is deleted");
+        return back()->with("info", "Unauthorized to delete this article");
     }
 }
